requirements.txt
setuptools<81
openai
openai-whisper
torch
torchaudio
torchvision
librosa
numpy
PyYAML
ffmpeg-python
opencv-python
pysrt
scenedetect
